hw3: сделаны правки

+++ src/main/java/ru/valerii/NauJava/entity/Transaction.java
@@ -3,11 +3,13 @@ package ru.valerii.NauJava.entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Transaction {
     private Long id;
-    private Double amount;
+    private BigDecimal amount;
     private Currency currency;
     private String description;
 }
